pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "saiql"
version = "0.2.0-alpha"
description = "Semantic Artificial Intelligence Query Language - AI-Native Database Engine"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Open Lore License v1.0"}
authors = [
    {name = "Apollo Raines", email = "apollo@saiql.ai"}
]
maintainers = [
    {name = "Apollo Raines", email = "apollo@saiql.ai"}
]
keywords = [
    "database",
    "ai",
    "machine-learning",
    "vector-search",
    "semantic-search",
    "lsm",
    "qipi",
    "loretoken"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Topic :: Database",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: Other/Proprietary License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
]

dependencies = [
    "numpy>=1.24.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "black>=23.0.0",
]
server = [
    "flask>=3.0.0",
    "gunicorn>=21.0.0",
]
postgres = [
    "psycopg2-binary>=2.9.0",
]
all = [
    "saiql[dev,server,postgres]",
]

[project.urls]
Homepage = "https://github.com/apolloraines/SAIQL"
Documentation = "https://github.com/apolloraines/SAIQL/tree/main/docs"
Repository = "https://github.com/apolloraines/SAIQL"
Issues = "https://github.com/apolloraines/SAIQL/issues"
Changelog = "https://github.com/apolloraines/SAIQL/blob/main/RELEASE_NOTES.md"

[project.scripts]
saiql = "bin.saiql:main"
saiql-server = "saiql_production_server:main"

[tool.setuptools]
packages = ["core", "storage", "interface", "security", "utils", "config", "shell", "bin"]
py-modules = ["saiql"]

[tool.setuptools.package-data]
"*" = ["*.json", "*.yaml", "*.yml"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=core --cov=storage --cov-report=term-missing"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
files = ["core", "storage"]

[tool.ruff]
line-length = 120
target-version = "py311"
select = ["E", "F", "W", "I", "N", "UP"]
ignore = ["E501"]  # Line too long (handled by formatter)

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # Unused imports in __init__.py

[tool.black]
line-length = 120
target-version = ["py311"]
include = '\.pyi?$'
